Seperate all code from extensions into smaller files.
extension.js
This is still the main file of the extension.
This has been condenced to essentially just the activate function with everything else abstracted away.

commands.js
Here we defined the actual funtionality of the 3 commands.

statusBarItems.js
This file defines a custom CustomStatusBarItem and creates items for the 3 icons,

utils.js
This file contains random helper or utility functions that are required